design: 불필요한 div 삭제 (#4)
